Clean-up, unused file removal, organize folders
