+++ backend/src/routes/index.js
@@ -1,12 +1,13 @@
-import { signUpRoute } from "./signUpRoute.js";
-import { loginRoute } from "./loginRoute.js";
+import { signUpRoute } from "./authRoutes/signUpRoute.js";
+import { resetPasswordRoute } from "./authRoutes/resetPasswordRoute.js";
+import { loginRoute } from "./authRoutes/loginRoute.js";
+import { verifyEmailRoute } from "./authRoutes/verifyEmailRoute.js";
+import { forgotPasswordRoute } from "./authRoutes/forgotPasswordRoute.js";
+import { getGoogleOauthUrlRoute } from "./authRoutes/getGoogleOauthUrlRoute.js";
+import { googleOauthCallbackRoute } from "./authRoutes/googleOauthCallbackRoute.js";
 import { testRoute } from "./testRoute.js";
 import { updateUserInfoRoute } from "./updateUserInfoRoute.js";
 import { testEmailRoute } from "./testEmailRoute.js";
-import { verifyEmailRoute } from "./verifyEmailRoute.js";
-import { forgotPasswordRoute } from "./forgotPasswordRoute.js";
-import { getGoogleOauthUrlRoute } from "./getGoogleOauthURLRoute.js";
-import { googleOauthCallbackRoute } from "./googleOauthCallbackRoute.js";
 import { getAllFeatures, insertFeatures } from "./featuresAndAppliances.js";
 import { addProperty } from "./addPropertyRoute.js";
 import { getAllUsersRoute } from "./getAllUsersRoute.js";
@@ -14,7 +15,6 @@ import { deleteUserRoute } from "./deleteUserRoute.js";
 import { getAllPropertiesRoute } from "./getAllPropertiesRoute.js";
 import { getAllAppliancesRoute } from "./getAllAppliancesRoute.js";
 import { getAllTasksRoute } from "./getAllTasksRoute.js";
-import { resetPasswordRoute } from "./resetPasswordRoute.js";
 
 export const routes = [
   addProperty,
