+++ frontend/src/pages/auth/EmailVerificationFail.js
