+++ frontend/src/pages/auth/EmailVerificationLandingPage.js
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
 import axios from "axios";
-import { useToken } from "../auth/useToken";
+import { useToken } from "../../auth/useToken";
 import { EmailVerificationSuccess } from "./EmailVerificationSuccess";
 import { EmailVerificationFail } from "./EmailVerificationFail";
